Commit 26

+++ IdleClicker_UnityBuild/Assets/Scripts/Battle/OpponentAnimationTrigger.cs
@@ -47,7 +47,7 @@ namespace IdleClicker.Battle
         public void SpawnPopupText()
         {
             PopupText popupTextInstance = Instantiate(popupText, canvas.transform);
-            popupTextInstance.SetValue(gainStrength);
+            popupTextInstance.SetValue((int)gainStrength);
 
             Vector2 spawnPosition = spawnPopupTextPoint.localPosition;
             spawnPosition.x += Random.Range(-spawnPopupTextCircleRadius, spawnPopupTextCircleRadius);
